% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImbTree.R
\name{ImbTree}
\alias{ImbTree}
\title{Title}
\usage{
ImbTree(
  Y_name,
  X_names,
  data,
  depth = 5,
  level_positive = "1",
  min_obs = 5,
  pct_obs = 0.1,
  type = "AUCg2",
  qval = 1,
  cp = 0,
  n_cores = 1,
  weights = NULL,
  AUC_weight = "none",
  cost = NULL,
  Class_threshold = "equal"
)
}
\arguments{
\item{Y_name}{Name of the target variable}

\item{X_names}{a}

\item{data}{a}

\item{depth}{a}

\item{level_positive}{a}

\item{min_obs}{a}

\item{pct_obs}{a}

\item{type}{a}

\item{qval}{a}

\item{cp}{a}

\item{n_cores}{a}

\item{weights}{a}

\item{AUC_weight}{a}

\item{cost}{a}

\item{Class_threshold}{a}
}
\value{

}
\description{
Title
}
\examples{

\dontrun{
  ImbTree()
}
}
