% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AUC.R
\name{MulticlassAUC}
\alias{MulticlassAUC}
\title{Title}
\usage{
MulticlassAUC(
  response,
  predictor,
  levels = base::levels(as.factor(response)),
  percent = FALSE,
  direction = c("auto", "<", ">"),
  weights = NULL,
  AUC_weight = F,
  cost = NULL,
  ...
)
}
\arguments{
\item{response}{a}

\item{predictor}{a}

\item{levels}{a}

\item{percent}{a}

\item{direction}{a}

\item{weights}{a}

\item{AUC_weight}{a}

\item{cost}{a}

\item{...}{a}
}
\value{

}
\description{
Title
}
\examples{

}
