% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AUC.R
\name{TwoclassAUC}
\alias{TwoclassAUC}
\title{Title}
\usage{
TwoclassAUC(response, predictor, weights = NULL, cost = NULL)
}
\arguments{
\item{response}{a}

\item{predictor}{a}

\item{weights}{a}

\item{cost}{a}
}
\value{

}
\description{
Title
}
\examples{

}
